Clean remaining compiler warnings in Gaussian and sampling code

+++ piquassoboost/sampling/simulation_strategies/GaussianSimulationStrategyFast_wrapper.cpp
@@ -216,7 +216,7 @@ GaussianSimulationStrategyFast_wrapper_simulate(GaussianSimulationStrategyFast_w
     // preallocate Python list to hold the calculated samples
     PyObject* PySamples = PyTuple_New( (Py_ssize_t) samples.size() );
 
-    for ( int idx = 0; idx < samples.size(); idx++ ) {
+    for ( size_t idx = 0; idx < samples.size(); idx++ ) {
         // release the C++ array from the ownership of the calculated data
         samples[idx].set_owner(false);
